Concentrate pre checks on new class CLIPrechecker
